refactor: `NormalDepartures` => `Departures`
"Normal" is generally not a useful prefix, and this resolves an
inconsistency where the backend refers to this component as just
`departures`.
