+++ assets/src/apps/v2/bus_shelter.tsx
@@ -34,7 +34,7 @@ import TwoMedium from "Components/v2/bus_shelter/flex/two_medium";
 import Placeholder from "Components/v2/placeholder";
 import LinkFooter from "Components/v2/bus_shelter/link_footer";
 import NormalHeader from "Components/v2/lcd/normal_header";
-import NormalDepartures from "Components/v2/departures/normal_departures";
+import Departures from "Components/v2/departures";
 import LcdSubwayStatus from "Components/v2/subway_status/lcd_subway_status";
 
 import EvergreenContent from "Components/v2/evergreen_content";
@@ -60,7 +60,7 @@ const TYPE_TO_COMPONENT = {
   placeholder: Placeholder,
   link_footer: LinkFooter,
   normal_header: NormalHeader,
-  departures: NormalDepartures,
+  departures: Departures,
   subway_status: LcdSubwayStatus,
   alert: FlexZoneAlert,
   full_body_alert: FullBodyAlert,
