+++ assets/src/apps/v2/gl_eink.tsx
@@ -22,7 +22,7 @@ import OneMedium from "Components/v2/eink/flex/one_medium";
 import Placeholder from "Components/v2/placeholder";
 import FareInfoFooter from "Components/v2/eink/fare_info_footer";
 import NormalHeader from "Components/v2/eink/gl_normal_header";
-import NormalDepartures from "Components/v2/departures/normal_departures";
+import Departures from "Components/v2/departures";
 import LineMap from "Components/v2/gl_eink_double/line_map";
 import EvergreenContent from "Components/v2/evergreen_content";
 import NoData from "Components/v2/eink/no_data";
@@ -59,7 +59,7 @@ const TYPE_TO_COMPONENT = {
   placeholder: Placeholder,
   fare_info_footer: FareInfoFooter,
   normal_header: NormalHeader,
-  departures: NormalDepartures,
+  departures: Departures,
   alert: MediumFlexAlert,
   full_body_alert: FullBodyTopScreenAlert,
   line_map: LineMap,
